add new header to all stacks with conditional rendering of back button and menu button, looks solid on client and pilot sides
