--- components/shared/NestedHeader.js
@@ -1,30 +1,0 @@
-import React from "react";
-import {
-  View,
-  StyleSheet,
-  Image,
-  Platform,
-} from "react-native";
-import { Ionicons } from "@expo/vector-icons";
-import hovrtekLogo from "../../assets/hovrtek_logo.png";
-
-const NestedHeader = () => {
-  return (
-    <View style={styles.wrapper}>
-      <Image source={hovrtekLogo} style={styles.hovrtekLogo} />
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  wrapper: {
-    justifyContent: "center"
-  },
-  hovrtekLogo: {
-    width: 160,
-    height: 30,
-    marginLeft: 30
-  }
-});
-
-export default NestedHeader;
