+++ navigation/ClientProfileNavigation.js
@@ -2,10 +2,8 @@ import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import ClientProfileScreen from "../screens/client/ClientProfileScreen";
 import ClientEditProfileScreen from "../screens/client/ClientEditProfileScreen";
-import MainHeader from '../components/shared/MainHeader';
-import NestedHeader from '../components/shared/NestedHeader';
+import GlobalHeader from "../components/shared/GlobalHeader";
 import { Ionicons } from "@expo/vector-icons";
-
 
 const ClientProfileNavigation = createStackNavigator();
 
@@ -18,7 +16,7 @@ function ClientProfileNavigator() {
         options={{
           animationEnabled: false,
           title: "Profile",
-          headerTitle: () => <MainHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
@@ -30,9 +28,10 @@ function ClientProfileNavigator() {
         name="ClientEditProfileScreen"
         component={ClientEditProfileScreen}
         options={{
+          headerLeft: null,
           animationEnabled: false,
           title: "Edit Profile",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100
