+++ navigation/PilotHomeStackNavigator.js
@@ -7,8 +7,8 @@ import ClientProfileScreen from "../screens/client/ClientProfileScreen";
 import ChatScreen from "../screens/messaging/ChatScreen";
 import PilotCreateProfileNavigation from "./PilotCreateProfileNavigation";
 import MyJobsScreen from "../screens/pilot/MyJobsScreen.js";
-import MainHeader from "../components/shared/MainHeader";
-import NestedHeader from "../components/shared/NestedHeader";
+import GlobalHeader from "../components/shared/GlobalHeader";
+
 import { Ionicons } from "@expo/vector-icons";
 
 const HomeStack = createStackNavigator();
@@ -21,7 +21,7 @@ const PilotHomeStackNavigator = () => {
         component={JobListScreen}
         options={{
           title: " ",
-          headerTitle: () => <MainHeader />,
+          headerTitle: () => <GlobalHeader isHome={true} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
@@ -32,98 +32,82 @@ const PilotHomeStackNavigator = () => {
         name="JobDetailsScreen"
         component={JobDetailsScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
-          },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
+          },   
         }}
       />
       <HomeStack.Screen
         name="AcceptJobScreen"
         component={AcceptJobScreen}
         options={{
+          headerLeft: null,
           title: "Accept Job",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
       <HomeStack.Screen
         name="ClientProfileScreen"
         component={ClientProfileScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
       <HomeStack.Screen
         name="ChatScreen"
         component={ChatScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
       <HomeStack.Screen
         name="MyJobsScreen"
         component={MyJobsScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <MainHeader />,
+          headerTitle: () => <GlobalHeader isHome={true} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"white"} />
-          ),
         }}
       />
       <HomeStack.Screen
         name="Profile"
         component={PilotCreateProfileNavigation}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <MainHeader />,
-          headerStyle: {
-            backgroundColor: "#092455",
-            height: 100,
-          },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"white"} />
-          ),
+          headerShown: false,
         }}
       />
     </HomeStack.Navigator>
