+++ navigation/PilotJobsStackNavigator.js
@@ -1,14 +1,10 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
-import JobListScreen from "../screens/pilot/JobListScreen";
 import JobDetailsScreen from "../screens/pilot/JobDetailsScreen";
-import AcceptJobScreen from "../screens/pilot/AcceptJobScreen";
 import ClientProfileScreen from "../screens/client/ClientProfileScreen";
 import ChatScreen from "../screens/messaging/ChatScreen";
-import PilotCreateProfileNavigation from "./PilotCreateProfileNavigation";
 import MyJobsScreen from "../screens/pilot/MyJobsScreen.js";
-import MainHeader from "../components/shared/MainHeader";
-import NestedHeader from "../components/shared/NestedHeader";
+import GlobalHeader from "../components/shared/GlobalHeader";
 import { Ionicons } from "@expo/vector-icons";
 
 const JobsStack = createStackNavigator();
@@ -21,65 +17,56 @@ const PilotJobsStackNavigator = () => {
         component={MyJobsScreen}
         options={{
           title: " ",
-          headerTitle: () => <MainHeader />,
+          headerTitle: () => <GlobalHeader isHome={true} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"white"} />
-          ),
         }}
       />
       <JobsStack.Screen
         name="JobDetailsScreen"
         component={JobDetailsScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
       <JobsStack.Screen
         name="ClientProfileScreen"
         component={ClientProfileScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
       <JobsStack.Screen
         name="ChatScreen"
         component={ChatScreen}
         options={{
+          headerLeft: null,
           title: " ",
-          headerTitle: () => <NestedHeader />,
+          headerTitle: () => <GlobalHeader isHome={false} />,
           headerStyle: {
             backgroundColor: "#092455",
             height: 100,
             borderBottomColor: "grey",
             borderBottomWidth: 10,
           },
-          headerBackImage: () => (
-            <Ionicons name="ios-arrow-round-back" size={50} color={"#092455"} />
-          ),
         }}
       />
     </JobsStack.Navigator>
