+++ navigation/PilotTabs.js
@@ -3,7 +3,6 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import PilotHomeStackNavigator from "./PilotHomeStackNavigator";
 import PilotJobsStackNavigator from "./PilotJobsStackNavigator";
 import MessagingNavigation from './MessagingNavigation';
-import MyJobsScreen from '../screens/pilot/MyJobsScreen';
 import { FontAwesome } from '@expo/vector-icons';
 import { Entypo } from '@expo/vector-icons';
 import { FontAwesome5 } from '@expo/vector-icons';
